Move frequency() tests into own file
As they are no longer part of the logical_meter.
